update from master

+++ gerapy/server/core/views.py
@@ -1,6 +1,5 @@
 import json, os, requests, time, pytz, pymongo, string
 from shutil import move, copy, rmtree
-from scrapyd_api import ScrapydAPI
 from requests.exceptions import ConnectionError
 from os.path import join, exists
 from django.shortcuts import render
@@ -14,7 +13,7 @@ from gerapy.server.server.settings import TIME_ZONE
 from gerapy.server.core.models import Client, Project, Deploy, Monitor
 from gerapy.server.core.build import build_project, find_egg
 from gerapy.server.core.utils import IGNORES, is_valid_name, copy_tree, TEMPLATES_DIR, TEMPLATES_TO_RENDER, \
-    render_template, get_traceback, scrapyd_url, log_url, get_tree
+    render_template, get_traceback, scrapyd_url, log_url, get_tree, get_scrapyd
 
 
 def index(request):
@@ -140,7 +139,7 @@ def spider_list(request, client_id, project_name):
     """
     if request.method == 'GET':
         client = Client.objects.get(id=client_id)
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         try:
             spiders = scrapyd.list_spiders(project_name)
             spiders = [{'name': spider, 'id': index + 1} for index, spider in enumerate(spiders)]
@@ -160,7 +159,7 @@ def spider_start(request, client_id, project_name, spider_name):
     """
     if request.method == 'GET':
         client = Client.objects.get(id=client_id)
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         try:
             job = scrapyd.schedule(project_name, spider_name)
             return JsonResponse({'job': job})
@@ -177,7 +176,7 @@ def project_list(request, client_id):
     """
     if request.method == 'GET':
         client = Client.objects.get(id=client_id)
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         try:
             projects = scrapyd.list_projects()
             return JsonResponse(projects)
@@ -263,13 +262,18 @@ def project_remove(request, project_name):
     :return: result of remove
     """
     if request.method == 'POST':
+        # delete deployments
+        project = Project.objects.get(name=project_name)
+        Deploy.objects.filter(project=project).delete()
+        # delete project
+        result = Project.objects.filter(name=project_name).delete()
+        # get project path
         path = join(os.path.abspath(os.getcwd()), PROJECTS_FOLDER)
         project_path = join(path, project_name)
         # delete project file tree
-        rmtree(project_path)
-        # delete project
-        result = Project.objects.filter(name=project_name).delete()
-        return JsonResponse({'result': result})
+        if exists(project_path):
+            rmtree(project_path)
+        return JsonResponse({'result': '1'})
 
 
 def project_version(request, client_id, project_name):
@@ -284,7 +288,7 @@ def project_version(request, client_id, project_name):
         # get client and project model
         client = Client.objects.get(id=client_id)
         project = Project.objects.get(name=project_name)
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         # if deploy info exists in db, return it
         if Deploy.objects.filter(client=client, project=project):
             deploy = Deploy.objects.get(client=client, project=project)
@@ -318,12 +322,14 @@ def project_deploy(request, client_id, project_name):
         project_path = join(path, project_name)
         # find egg file
         egg = find_egg(project_path)
+        if not egg:
+            return JsonResponse({'message': 'egg not found'}, status=500)
         egg_file = open(join(project_path, egg), 'rb')
         # get client and project model
         client = Client.objects.get(id=client_id)
         project = Project.objects.get(name=project_name)
         # execute deploy operation
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         try:
             scrapyd.add_version(project_name, int(time.time()), egg_file.read())
             # update deploy info
@@ -375,6 +381,8 @@ def project_build(request, project_name):
         description = data['description']
         build_project(project_name)
         egg = find_egg(project_path)
+        if not egg:
+            return JsonResponse({'message': 'egg not found'}, status=500)
         # update built_at info
         built_at = timezone.now()
         # if project does not exists in db, create it
@@ -403,8 +411,9 @@ def project_generate(request, project_name):
     if request.method == 'POST':
         # get configuration
         configuration = Project.objects.get(name=project_name).configuration
+        if not configuration:
+            return JsonResponse({'message': 'Invalid configuration'}, status=500)
         configuration = json.loads(configuration)
-        
         if not is_valid_name(project_name):
             return JsonResponse({'message': 'Invalid project name'}, status=500)
         # remove original project dir
@@ -519,7 +528,7 @@ def job_list(request, client_id, project_name):
     """
     if request.method == 'GET':
         client = Client.objects.get(id=client_id)
-        scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+        scrapyd = get_scrapyd(client)
         try:
             result = scrapyd.list_jobs(project_name)
             jobs = []
@@ -551,7 +560,9 @@ def job_log(request, client_id, project_name, spider_name, job_id):
             # get last 1000 bytes of log
             response = requests.get(url, timeout=5, headers={
                 'Range': 'bytes=-1000'
-            })
+            }, auth=(client.username, client.password) if client.auth else None)
+            # change encoding
+            response.encoding = response.apparent_encoding
             # log not found
             if response.status_code == 404:
                 return JsonResponse({'message': 'Log Not Found'}, status=404)
@@ -573,7 +584,7 @@ def job_cancel(request, client_id, project_name, job_id):
     if request.method == 'GET':
         client = Client.objects.get(id=client_id)
         try:
-            scrapyd = ScrapydAPI(scrapyd_url(client.ip, client.port))
+            scrapyd = get_scrapyd(client)
             result = scrapyd.cancel(project_name, job_id)
             return JsonResponse(result)
         except ConnectionError:
